Updated my ORDER BY to id from date

+++ src/components/AdminItems/AdminItems.jsx
@@ -9,21 +9,28 @@ import axios from 'axios';
 function AdminItems({ item, getSurveyResults }) {
     //this function will allow me to make a delete request to the server using item.id
   const onDeleteButton = () => {
-    let deleteId = item.id;
-    console.log("on Delete Button", deleteId);
-    //making request for deletion 
-    axios({
-      method: "DELETE",
-      url: `/api/survey/${deleteId}`, //passing the item id 
-    })
-      .then((response) => {
-        console.log("DELETE /api/survey/ response is:", response);
-        getSurveyResults(); // will refresh page with current data from my db
-      })
-      .catch((error) => { // will catch error 
-        //this will log the error to the console.
-        console.log("DELETE /api/survey/ error is:", error);
-      });
+  
+    if (confirm("Are You Sure?")) {
+          let deleteId = item.id;
+          console.log("on Delete Button", deleteId);
+          //making request for deletion
+          axios({
+            method: "DELETE",
+            url: `/api/survey/${deleteId}`, //passing the item id
+          })
+            .then((response) => {
+              console.log("DELETE /api/survey/ response is:", response);
+              getSurveyResults(); // will refresh page with current data from my db
+            })
+            .catch((error) => {
+              // will catch error
+              //this will log the error to the console.
+              console.log("DELETE /api/survey/ error is:", error);
+            });
+   
+    } else {
+      alert("Delete request cancelled.");
+    }
   };
   return (
     <>
